backend/requirements.txt
python-dotenv
agno
openai
groq
llama-index-llms-groq
